Removed unnessesary print
The .unsupported case is by no means a problem - simply using postgres enums will cause each and every bind to print out this message. As the function still returns the data (in the form of bytes) it makes more sense that any issue with this will be caught in higher level testing.